requirements.txt
python-telegram-bot==13.15
Flask==2.2.5
python-dotenv==1.0.1